Modularise UI code
